requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.1.2
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
GDAL==3.4.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.6.0
